double to float

+++ ShapeArray.h
@@ -17,22 +17,22 @@ struct shpmainheader
 	int version;
 	int shapetype;
 	int pad;
-	double xmin, ymin;
-	double xmax, ymax;
-	double zmin, zmax;
-	double mmin, mmax;
+	float xmin, ymin;
+	float xmax, ymax;
+	float zmin, zmax;
+	float mmin, mmax;
 };
 
 //pack xyz components into a point
 struct point3d
 {
-	double x, y, z;
+	float x, y, z;
 };
 
 //header for shapes
 struct shpheader
 {
-	double xmin, ymin, xmax, ymax, zmin, zmax;
+	float xmin, ymin, xmax, ymax, zmin, zmax;
 	int numparts;
 	int numpoints;
 };
